teach internals of reference counted string impl HTML-like strings
This change discriminates between regular and HTML-like strings within the
reference counted strings code. This is related to #2089, but has no impact on
it yet because the comparison and hash functions do not interpret this
information yet.

Note that this makes the semantics of some API functions a little weird. In
particular, `agstrbind` and `agstrfree` take no parameter indicating whether the
input string is a regular string or an HTML-like string. So these now attempt
their operation on a regular string and then, if that fails, attempt it on an
HTML-like string. This seems like the best (least surprising) thing we can do
within the constraints of the current API. In future we may want to improve the
API in this area.