agcopyattr: fix: create HTML-like strings as HTML-like to begin with
This fixes a latent issue where HTML-like labels were copied by creating the
label as a regular string and then subsequently marking it as an HTML-like
string. One problem with this is that the string interning implementation
de-dupes this initial string creation against existing (non-HTML-like) strings.
As a result it is not possible to add a regular string and an HTML-like string
with the same content (#2089).

If we were to fix #2089, it would still not be possible to faithfully copy a set
of attributes because of the above described way `agcopyattr` initially creates
strings as regular strings. That is, it would get confused about the HTML-ness
of two strings with the same content. This change fixes this problem in
preparation for dealing with #2089.