remove agmarkhtmlstr
Upcoming changes will make the HTML-ness of a string considered as part of
comparison and hashing of reference counted strings. This means it will no
longer be possible to toggle the HTML-ness of a string after creation. That is,
inserting a reference counted string and *then* setting the `is_html` bit will
alter its hash, making it unfindable in future. Related to #2089.