flip Smyrna default to disabled in the Autotools build system
This is step 1 of a 4 step plan to remove Smyrna:¹

  1. Default it to disabled (this commit).
  2. Disable it in CI jobs that build release artifacts.
  3. Remove build system support.
  4. Remove Smyrna code.

The current plan is to have ≥1 release in between each of these steps. This
should make backing out these steps straightforward if compelling reasons for
continuing to support Smyrna emerge.

The present change only affects users who build Smyrna using the Autotools build
system on Linux. Other environments are unaffected:

  - CMake build system: no support for building Smyrna (#1836).
  - macOS: Smyrna is disabled (#1854).
  - Windows: Smyrna is built but seems unusable (#1829).

Related, Smyrna is also not currently built on CentOS (#1834).

¹ https://gitlab.com/graphviz/graphviz/-/issues/1834#note_576332436