CMake: FindLTDL: remove useless misspelled name of libltdl
The name libtldl was clearly a misspelling of libltdl introduced
already when this file was created in commit
3b567c4bd938e2e70478b22d5efae98bedf9ed8f.

There's however no need to specify anything more then the ltdl name
itself since CMake does this kind of magic automatically. From
https://cmake.org/cmake/help/latest/command/find_library.html:

"Each library name given to the NAMES option is first considered as a
library file name and then considered with platform-specific prefixes
(e.g. lib) and suffixes (e.g. .so)."