gvc: CMake: fix: link to the ltdl lib only if we have it
Before this we tried to use the ltdl link library when not on Windows,
regardless of whether we found it or not. On Windows, we currently
only use the ltdl compatibility header introduced in
d70cdba7937724c19393d97bd280320cdf8c3c8f, not the link library itself.

Fixes the LTDL_LIBRARY part of the following error on non-Windows
systems when we don't have ltdl:

CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
/home/magjac/graphviz/lib/gvc/LTDL_INCLUDE_DIR
   used as include directory in directory /home/magjac/graphviz/lib/gvc
LTDL_LIBRARY (ADVANCED)
    linked by target "gvc" in directory /home/magjac/graphviz/lib/gvc

There are other problems when the LTDL library is not available that
will be fixed in upcoming commits, hence no changelog entry yet.