remove Exshort_t type
This type discrimination did not seem to make much sense (why select between int
and short based on the sizes of long and int?), but more problematically was not
working as intended. This header does not include limits.h, so both LONG_MAX and
INT_MAX were undefined and this expression defaulted to 0 regardless of the
sizes of these types.

Closes #1799.