don't read the VERSION file for CMake builds in ci/build.sh
The CMake build system generates this file itself and nothing in this
script uses it for the CMake builds.

An upcoming commit will change the CI configuration to start the CMake
builds jobs directly without waiting for the portable-source job and
since that job produces the VERSION file, it will no longer be present
when the CMake builds jobs start.