fix multiple layout test case using rvalue refs
The first layout of a graph must be destroyed before a second layout
is created.

This failure went undetected in CI since no test programs ran, which
in turn was caused by not archiving the test programs or the
CTestTestfile.cmake files in the preceding build job.

The error message was:

-------------------------------------------------------------------------------
Multiple layouts of the same graph can use different contexts passed as rvalue
refs
-------------------------------------------------------------------------------
../tests/test_GVLayout_construction.cpp:139
...............................................................................

../tests/test_GVLayout_construction.cpp:139: FAILED:
due to unexpected exception with message:
  Previous layout not yet destroyed