exsub: remove use of sfio temporary buffer
This removes yet another use of `ex->tmp` within exeval.c code. It is replaced
with a dynamically allocated buffer from vmalloc. This not only reduces reliance
on sfio, but is likely more efficient as data is written directly into a vmalloc
buffer rather than being written to an sfio buffer and then copied to a vmalloc
buffer. Related to #1998.