CMake: add with_zlib option
The zlib library is already optional. This option allows the user to
explictly not use it even if they have it.

This will be used in an upcoming commit to add a minimal CMake build
job to CI which will test that it is possible to build Graphviz
without any of the optional libraries.