gvc: CMake: fix: use the ltdl include dir only if we have it
Fixes this error when we don't have ltdl:

CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
C:/Users/magja/graphviz/lib/gvc/LTDL_INCLUDE_DIR
   used as include directory in directory C:/Users/magja/graphviz/lib/gvc

On Windows, we currently only use the ltdl compatibility header
introduced in d70cdba7937724c19393d97bd280320cdf8c3c8f, not the link
library itself. Hence the WIN32 OR MINGW guard for the link library,
but not for the include directory.

There are other problems when the LTDL library is not available that
will be fixed in upcoming commits, hence no changelog entry yet.