fix dangerous-default-value issues in gvtest.py test suite support
Pylint warns that using `[]` as a default is dangerous due to modifications to
the parameter unintentionally persisting across calls. I do not believe that
happens in this case, but it is better to rephrase this code to pacify Pylint
and make it clearer to readers that this is not happening.