remove `make check` Python targets
The current Graphviz test suite is orchestrated via Pytest and assumes it is
dealing with an installed Graphviz (e.g. headers and libraries available in
expected locations). The legacy Autotools check targets gave the incorrect
impression that this was the current test suite and it was expected to pass. It
has not been possible to run these legacy test scripts for a long time.

This commit removes the legacy testing pieces to avoid future confusion. In
future it may be useful to plumb the current Pytest test suite to the Autotools
post-install test targets (`make installcheck`).

Closes #2112.