ci/build.sh: remove obsolete workaround for missing ID_LIKE in /etc/os-release
This workaround was introduced in
029acf11d1d9a9eb20eddae84cadb61447da95f3, but was actually never
brought into use when 9ca0ccd4e368d0e9bd545e5f1698b8eef931dbbc
introduced an ASan build to CI, which used a Ubunutu docker container
instead of the native GitLab Debian Linux runner for which the
workaround was intended.

Also Fedora doesn't have ID_LIKE, but the workaround was really never
necessary for it either.