Ensure correct file-level dependency for generated file in cmake generated projects
When using ninja parallel compilation, it is possible that the generated ninja
file might not create the necessary files, color_lib and svgcolor_lib, before
they are needed and this results in a failed build. Providing an absolute path
to depend on in the add_custom_command rectifies the situation.

Closes #2119