ci/build.sh: fix missing VERSION_ID on MSYS2/MinGW
This change will allow treating unset variables as errors in an
upcoming commit.

This problem also caused the build artifacts to be moved to the wrong
directory, but that did not have any consequences since the build
artifacts are currently not archived in the CI jobs. An upcoming
commit in this series will however change that.