gvmap: autotools: link to the gvc lib instead of the static common lib
The common lib is part of gvc and should not be linked to directly.

This will allow an upcoming MR to correct the storage class attributes
in lib/common/globals.h.

Wihout this commit, many errors like this would appear when linking
gvmap:

/usr/lib/gcc/x86_64-pc-cygwin/10/../../../../x86_64-pc-cygwin/bin/ld: ../../lib/edgepaint/.libs/libedgepaint_C.a(node_distinct_coloring.o): in function `node_distinct_coloring_internal2':
/home/magja/graphviz/lib/edgepaint/node_distinct_coloring.c:137: undefined reference to `__imp_Verbose'

Towards https://gitlab.com/graphviz/graphviz/-/issues/2058.