fix inability to include and use gvc.h from a parent CMake project
When attempting to declare a dependency on gvc when using Graphviz as a CMake
subproject, the include path necessary to compile gvc headers would not be
propagated to dependent targets. The result would be a compilation error when
the transitive includes of gvc.h could not be found.

To validate this change, create an empty directory and add the following
CMakeLists.txt:

  project(foo)
  add_subdirectory(graphviz)
  add_executable(bar main.c)
  target_link_libraries(bar gvc)

Add the following main.c:

  #include <gvc/gvc.h>
  int main(void) { return 0; }

Clone Graphviz into the subdirectory 'graphviz'. Then the standard CMake steps:

  mkdir build
  cd build
  cmake ..
  make

Note that Graphviz dependencies and then 'bar' are correctly built.

I am not a CMake expert; this was written based on my best guess of what
Graphviz should be doing according to the CMake docs. Assuming this is correct,
other Graphviz targets probably need similar fixes.

Fixes #1477, #2109.